Cleaning of the constructor of AdBlockRule
Basic cleaning and bug fixing of AdBlockRule.

The options shoud be parsed before checking if the filter is a regexp,
otherwhise a regexp with option will never match the rule.
For example:
-"/.*/$xmlhttprequest" is a valid regexp that would not be matched.
-"/.*$xmlhttprequest/" is a valid regexp without options and the option
should not have been parsed.

The matching of the option should match the last index of "$", not the
first, for the same reason as above.

Use mid() to split the vector at once instead QString::mid() +
QString::lef().

Clean the coding style to follow the conventions of KDE.