Split AdBlock rule in two classes to move the implementation out of it
In order to make special matching rules, we need specialization of the
implementation depending on the type of rule.

The previous AdBlockRule was entierly based on regexp. The new one
is only a factory to a AdBlockRuleImpl, and delegate everything to
this implementation.

This will allow faster specialization of the ad block rules in the
future.