Add support for domain option in the ad block module
Some ads were not filtered because they were incorrectly matched by
the whitelist. This is because we ignore options, including the domain
restrictions.

For example, the white filter:
@@||pagead2.googlesyndication.com/pagead/show_ads.js$domain=kde.org
would match any page regardless of the domain restriction. So no ads
from pagead2.googlesyndication.com were filtered.

This patch adds support for "domain" options in the fallback rules
(the other AdBlockRuleImpl rejects options). The domain of the frame
making the request is now compared to the option to take the right
decision.

This patch requires Qt 4.7 with a recent QtWebKit 2.0.