Make the FindBar a bit more robust
The FindBar was slots show() and hide() instead of the virtual method
QWidget::setVisible(). This can lead to subtle bugs, because the
behavior is different depending on the type of call (signal-slot
connection VS direct call on a pointer to QWidget) and the function
called (hide() is != than setVisible(false)).