User Agent Switcher.
This patch should implement the UA switcher ability for rekonq.
This is a first implementation, but it seems working quite well.
It is based on KDE UA management and should be fully compatible with konqueror's one i.e.
it should be possible using alternatively rekonq and konqueror sharing the same setting.

Anyway, this is a different implementation from the konqueror's one. Simpler and based on the idea
of a future moving to a plugin.