Add bookmarks by dropping inside the bk toolbar.
This patch, based on Furkan's work on this, is quite different from the original for the
following reasons:
first, the "need" to not modify a lot actual code in the urlbar
(disabling dragging and reenabling by code. What are the implications? Who really knows this?)
Second, the problem of showing "as urls" (showing their icon) TEXT mimeData: not a good choice, IMHO.
Actual code is cleaner and works with every drop inside the bk toolbar
(In case of text, checking if the derivated url is valid)