fix input of "/" in url bar when using NEO keyboard layout
Using the NEO keyboard layout, you need a combination of keys to trigger
the slash (mod3+s). Unfortunatly this triggers the shortcut for
searching inside the document. This way the alternate shortcut for
searching needs to be removed.

The same bug existed in kopete: bug 267568

Thanks goes to Thomas Murach for helping me grepping the source code.
This is my first contribution in code to KDE. Hurray!