New History Manager, NO MORE QWebHistoryInterface based.
We are just saving datas on our own, I don't like QtWebKit APIs
to do it. We now call addHistoryEntry(url, title)
on loadFinished.
Data structure will remain the same as usual, but I think API
is better and data are now saved when we really have them
ready.

REVIEW:104257