Include Tab History when Cloning a Tab
Implemented through a new loadUrl method that allows for the caller to
specify the QWebHistory pointer from where to copy the history.
Redirected the original loadUrl slot to use this new method internally.
The method returns the WebTab * just in case we need to maninpulate the
created tab in the future.

REVIEW: 104082