"Workaround" cookie management in private browsing mode
From the note in webpage file:
I'm sorry I really cannot let KIO work as needed in private browsing
mode. The problem is basically cookie related. This way we lose some
features in private browsing mode, but we ensure PRIVACY!

This change cannot be reverted until a proper fix for KIO
(or the right workaround for rekonq) will be found.

BUG: 303552