honor filetype setting WRT embedding
there are 3 distinct states a filetype can have WRT kpart embeding
- always embed
- never embed
- do whatever the parent node does (e.g. application/foo would take the
  setting of application)

since the logic to determine which of those it is going to be we are using
a bit of code imported from konqueror deciding in a boolean fashion
whether or not we are supposed to embed or not. this is particularly non-
intrusive as the decision directly relates to whether a kpart is created,
if not the file will simply be krun'.

this change is using static functions for the imported code. rationale
being that they are in fact static and not having them reflected in the
header makes them all the easier to remove should a better solution
arise in the future.

with that in mind: while the code is copy'n'pastable it seems like a
good idea to move this into some shared library in the long term such
that konqueror and rekonq (and any other kpart enabled app) can use the
same code.

REVIEW: 109942
BUG: 240400
BUG: 279423