Adding domain-specific hiding support to the adblocker.
The domain-specific rules are stored in a QMultiHash, where the key
is the domain and the values are the rules for the specified domain.
This causes redundancy, because on rule can be applied to more than
one domain, but has a really fast look-up time.
The code for the generic hiding has not changed and is just relocated.

REVIEW: 111521
REVIEWED-BY: adjam