Implement Unique Mode Properly in Rekonq
The current patch just makes it possible to open several Rekonq
applications. It does not do the right thing when a Rekonq window is
already open in the current activity and a user clicks a link elsewhere
(step 4 in the Testing Done section) because it starts a brand new
Rekonq window, but that's a different patch. It also does some funky
thing asking you if you want to restore the previous session when
nothing has crashed

REVIEW: 120794